+++ src/modules/dish/dish.service.ts
@@ -10,7 +10,6 @@ import { UserDto } from '../user/user.dto';
 import { DishRepository } from '../../mongodb/repositories/dish.repository';
 import { MealType } from '../../common/enums';
 import { DetailedDish, DishRating, MergedSearchQueries, ProposedDish, RatedDish } from './dish.types';
-import { SpoonacularApiService } from '../api/spoonacular/spoonacular.api.service';
 import {
     getQueryWithIngredientsAndDishType,
     mergeSearchQueries,
@@ -30,6 +29,8 @@ import { DishRatingRepository } from '../../mongodb/repositories/dish-rating.rep
 import { DishRatingDocument } from '../../mongodb/documents/dish-rating.document';
 import { sortDescendingRelevance } from '../../common/helpers';
 import { ForbiddenException } from '../../exceptions/forbidden-exception';
+import { ExternalApiService } from '../api/external-api.service';
+import { getFulfilledPromiseResults } from '../../utils';
 
 @Injectable()
 export class DishService {
@@ -42,7 +43,7 @@ export class DishService {
         private redisService: RedisService,
         private loggerService: LoggerService,
         private ingredientService: IngredientService,
-        private spoonacularApiService: SpoonacularApiService
+        private externalApiService: ExternalApiService
     ) {}
 
     async create(createDishDto: CreateDishDto<DishIngredientWithoutImage>, user: UserDto): Promise<DishDocument> {
@@ -261,10 +262,7 @@ export class DishService {
             }
         }
 
-        // FIXME: Change to Promise.allSettled - Due to 402 Payment Required Error.
-        const datasets: Array<DetailedDish | null> = await Promise.all([
-            this.spoonacularApiService.getDishDetails(`recipes/${id}/information?apiKey=${process.env.SPOONACULAR_API_KEY}`, `recipes/${id}/analyzedInstructions?apiKey=${process.env.SPOONACULAR_API_KEY}`)
-        ]);
+        const datasets = await this.getDatasets<DetailedDish>(...this.externalApiService.getDishDetails(id));
 
         const filteredDish: DetailedDish = datasets.find(dish => dish !== null);
 
@@ -347,11 +345,7 @@ export class DishService {
 
         const ingredients = [...filteredIngredients, ...this.ingredientService.getAllPantryIngredients()];
 
-        // FIXME: Change to Promise.allSettled - Due to 402 Payment Required Error.
-        const datasets: Array<RatedDish[] | null> = await Promise.all([
-            this.dishRepository.getDishes(ingredients),
-            this.spoonacularApiService.getDishes(process.env.SPOONACULAR_API_KEY, 'recipes/findByIngredients', ingredients, type)
-        ]);
+        const datasets = await this.getDatasets(this.dishRepository.getDishes(ingredients), ...this.externalApiService.getDishes(ingredients, type));
 
         const dishes: RatedDish[] = datasets.flat().filter(dish => dish.relevance > 0).sort(sortDescendingRelevance);
         await this.redisService.saveDishResult('merged', query, dishes, 12 * HOUR);
@@ -367,11 +361,7 @@ export class DishService {
         const searchQueries: SearchQueryDocument[] = await this.searchQueryRepository.findAll({ date: { $gte: dateFilter }, login: user.login });
         const mergedSearchQueries: MergedSearchQueries = mergeSearchQueries(searchQueries);
         const ingredientsList = Object.keys(mergedSearchQueries);
-        // FIXME: Change to Promise.allSettled - Due to 402 Payment Required Error.
-        const datasets: Array<RatedDish[] | null> = await Promise.all([
-            this.dishRepository.getDishes(ingredientsList),
-            this.spoonacularApiService.getDishes(process.env.SPOONACULAR_API_KEY, 'recipes/findByIngredients', ingredientsList)
-        ]);
+        const datasets = await this.getDatasets(this.dishRepository.getDishes(ingredientsList), ...this.externalApiService.getDishes(ingredientsList));
         const dishes: RatedDish[] = datasets.flat().sort(sortDescendingRelevance);
         const proposedDishes: ProposedDish[] = proceedRatedDishesToProposedDishes(dishes, mergedSearchQueries);
 
@@ -417,10 +407,7 @@ export class DishService {
                 return true;
             }
         } else {
-            // FIXME: Change to Promise.allSettled - Due to 402 Payment Required Error.
-            const datasets: Array<DetailedDish | null> = await Promise.all([
-                this.spoonacularApiService.getDishDetails(`recipes/${dishId}/information?apiKey=${process.env.SPOONACULAR_API_KEY}`, `recipes/${dishId}/analyzedInstructions?apiKey=${process.env.SPOONACULAR_API_KEY}`)
-            ]);
+            const datasets = await this.getDatasets(...this.externalApiService.getDishDetails(dishId));
 
             const filteredDish: DetailedDish = datasets.find(dish => dish !== null);
 
@@ -516,5 +503,9 @@ export class DishService {
         this.loggerService.info(context, `Successfully added a new rating for "${createRatingBody.dishId}" dish by "${user}" user.`);
 
         return await this.dishRatingRepository.create({ ...createRatingBody, user, posted: Date.now() });
+    }
+
+    private getDatasets<T>(...datasets: Promise<T>[]): Promise<T[]> {
+        return getFulfilledPromiseResults<T>(datasets);
     }
 }