+++ src/utils.ts
@@ -15,3 +15,11 @@ export function getCorsOrigins(): string[] {
 
     return origins;
 }
+
+export async function getFulfilledPromiseResults<T>(promises: Promise<T>[]): Promise<T[]> {
+    const results = await Promise.allSettled<T>(promises);
+
+    return results
+        .filter(result => result.status === 'fulfilled')
+        .map(result => (result as PromiseFulfilledResult<T>).value);
+}